Move the ast package into the compiler package.
