+++ internal/vm/checker/checker.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/golang/glog"
 	"github.com/google/mtail/internal/metrics"
-	"github.com/google/mtail/internal/vm/ast"
+	"github.com/google/mtail/internal/vm/compiler/ast"
 	"github.com/google/mtail/internal/vm/errors"
 	"github.com/google/mtail/internal/vm/parser"
 	"github.com/google/mtail/internal/vm/symbol"
