+++ internal/vm/codegen/codegen.go
@@ -12,7 +12,7 @@ import (
 	"github.com/golang/glog"
 	"github.com/google/mtail/internal/metrics"
 	"github.com/google/mtail/internal/metrics/datum"
-	"github.com/google/mtail/internal/vm/ast"
+	"github.com/google/mtail/internal/vm/compiler/ast"
 	"github.com/google/mtail/internal/vm/code"
 	"github.com/google/mtail/internal/vm/errors"
 	"github.com/google/mtail/internal/vm/object"
