+++ internal/vm/compiler/ast/walk.go
