+++ internal/vm/parser/sexp.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 
 	"github.com/google/mtail/internal/metrics"
-	"github.com/google/mtail/internal/vm/ast"
+	"github.com/google/mtail/internal/vm/compiler/ast"
 	"github.com/google/mtail/internal/vm/symbol"
 )
